requirements.txt
gym~=0.19.0
sympy~=1.9
numpy~=1.22.0
pandas~=1.3.5
pytorch~=1.10.0
torch-geometric~=2.0.3
numexpr~=2.7.3
wolframclient~=1.1.7
multiprocess~=0.70.12.2
pathos~=0.2.8